refactor: expose compileMDX API
