+++ packages/gatsby-plugin-mdx/src/gatsby-node.ts
@@ -1,13 +1,12 @@
-import type { GatsbyNode, NodeInput, NodePluginArgs } from "gatsby"
+import type { GatsbyNode, NodeInput } from "gatsby"
 import type { FileSystemNode } from "gatsby-source-filesystem"
 import type { Options } from "rehype-infer-description-meta"
-import type { IFileNode, IMdxMetadata, IMdxNode, NodeMap } from "./types"
+import type { IFileNode, NodeMap } from "./types"
 
 import path from "path"
 import { sentenceCase } from "change-case"
 import fs from "fs-extra"
 import grayMatter from "gray-matter"
-import deepmerge from "deepmerge"
 
 import {
   defaultOptions,
@@ -15,7 +14,7 @@ import {
   IMdxPluginOptions,
 } from "./plugin-options"
 import { IGatsbyLayoutLoaderOptions } from "./gatsby-layout-loader"
-import compileMDX from "./compile-mdx"
+import { compileMDX, compileMDXWithCustomOptions } from "./compile-mdx"
 import { IGatsbyMDXLoaderOptions } from "./gatsby-mdx-loader"
 import remarkInferTocMeta from "./remark-infer-toc-meta"
 
@@ -143,58 +142,6 @@ export const preprocessSource: GatsbyNode["preprocessSource"] = async (
   )
 
   return processedMDX.toString()
-}
-
-/**
- * This helper function allows you to inject additional plugins and configuration into the MDX
- * compilation pipeline. Very useful to create your own resolvers that return custom metadata.
- * Internally used to generate the tables of contents and the excerpts.
- */
-export const compileMDXWithCustomOptions = async ({
-  pluginOptions,
-  customOptions,
-  getNode,
-  getNodesByType,
-  pathPrefix,
-  reporter,
-  cache,
-  mdxNode,
-}: {
-  pluginOptions: IMdxPluginOptions
-  customOptions: Partial<IMdxPluginOptions>
-  getNode: NodePluginArgs["getNode"]
-  getNodesByType: NodePluginArgs["getNodesByType"]
-  pathPrefix: string
-  reporter: NodePluginArgs["reporter"]
-  cache: NodePluginArgs["cache"]
-  mdxNode: IMdxNode
-}): Promise<{
-  processedMDX: string
-  metadata: IMdxMetadata
-} | null> => {
-  const customPluginOptions = deepmerge(
-    Object.assign({}, pluginOptions),
-    customOptions
-  )
-
-  // Prepare MDX compile
-  const mdxOptions = await enhanceMdxOptions(customPluginOptions, {
-    getNode,
-    getNodesByType,
-    pathPrefix,
-    reporter,
-    cache,
-  })
-  if (!mdxNode.parent) {
-    return null
-  }
-  const fileNode = getNode(mdxNode.parent)
-  if (!fileNode) {
-    return null
-  }
-
-  // Compile MDX and extract metadata
-  return compileMDX(mdxNode, fileNode, mdxOptions)
 }
 
 export const createSchemaCustomization: GatsbyNode["createSchemaCustomization"] =
