+++ packages/gatsby-plugin-mdx/src/index.ts
@@ -1,1 +1,1 @@
-// noop
+export * from "./compile-mdx"
